+++ CHANGELOG.md
@@ -7,3 +7,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 - Remove Get Help button on system packs
 - Open folder after exporting an instance
+- Show export button after adding/editing mods